+++ RPGCuzWhyNot/HealthSystem/IArmor.cs
@@ -1,6 +1,6 @@
 namespace StaterZ.Core.HealthSystem {
 	public interface IArmor {
-		float OnDamageModify(float amount);
-		float OnHealModify(float amount);
+		int OnDamageModify(int amount);
+		int OnHealModify(int amount);
 	}
 }
